Rename run event package
