+++ src/main/kotlin/org/coner/drs/ui/main/MainController.kt
@@ -1,7 +1,7 @@
 package org.coner.drs.ui.main
 
 import org.coner.drs.io.DrsIoController
-import org.coner.drs.ui.run_event.RunEventFragment
+import org.coner.drs.ui.runevent.RunEventFragment
 import org.coner.drs.ui.start.StartView
 import org.coner.drs.ui.choose_event.ChooseEventView
 import tornadofx.*
