+++ src/main/kotlin/org/coner/drs/ui/runevent/RegistrationHint.kt
@@ -1,3 +1,3 @@
-package org.coner.drs.ui.run_event
+package org.coner.drs.ui.runevent
 
 data class RegistrationHint(val category: String, val handicap: String, val number: String)
