+++ src/main/kotlin/org/coner/drs/ui/runevent/RunEventModel.kt
@@ -1,4 +1,4 @@
-package org.coner.drs.ui.run_event
+package org.coner.drs.ui.runevent
 
 import io.reactivex.disposables.CompositeDisposable
 import javafx.beans.property.SimpleObjectProperty
