+++ src/main/kotlin/org/coner/drs/ui/runevent/TimerConfigurationView.kt
@@ -1,4 +1,4 @@
-package org.coner.drs.ui.run_event
+package org.coner.drs.ui.runevent
 
 import javafx.scene.control.ButtonBar
 import org.coner.drs.TimerConfiguration
