+++ src/main/kotlin/org/coner/drs/ui/runevent/TimerView.kt
@@ -1,4 +1,4 @@
-package org.coner.drs.ui.run_event
+package org.coner.drs.ui.runevent
 
 import javafx.geometry.Orientation
 import javafx.stage.Modality
